Added event system for multiplicator and score. Created new event call with data. Deleted rayui_impl. Rewrote CMake.

+++ include/button.h
@@ -43,7 +43,7 @@ class tetrisButton {
         textureStyle _style;
         buttonState _state;
         std::string _text;
-        int _textSize;
+        float _textSize;
 };
 
 float tetrisButtongetTotalWidth(Size2 size);
